fix py lib
