+++ packages/py-moose-lib/setup.py
@@ -1,6 +1,8 @@
 import sys
 from setuptools import setup, find_packages
 
+version = "0.0.0"
+
 if '--version' in sys.argv:
     index = sys.argv.index('--version')
     sys.argv.pop(index)
